requirements.txt
absl-py==0.9.0
appdirs==1.4.3
appnope==0.1.0
attrs==19.3.0
autopep8==1.5
backcall==0.1.0
bleach==3.1.1
brewer2mpl==1.4.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
entrypoints==0.3
filelock==3.0.12
future==0.18.2
google-auth==1.11.3
google-auth-oauthlib==0.4.1
graphviz==0.14
grpcio==1.27.2
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.1
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
moviepy==1.0.1
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
proglog==0.1.9
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
PyYAML==5.3
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.2
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
torch==1.5.0
torchvision==0.6.0
tornado==6.0.3
tqdm==4.43.0
traitlets==4.3.3
urllib3==1.25.8
virtualenv==20.0.17
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
zipp==3.1.0
